users/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.2.0
Markdown==3.4.1
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
zipp==3.11.0